requirements.txt
Flask
python-dotenv
requests
Flask-SQLAlchemy